checkpoint
mostly moved stuff around and cleaned things up

added one or two error checks

shouldn't have combined these into one change but c'est la vie
